backend/requirements.txt
# Core FastAPI framework and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
psycopg2-binary==2.9.9
supabase==2.3.4

# Environment configuration
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Authentication and JWT
PyJWT==2.8.0

# AI and Machine Learning
transformers==4.36.2
torch==2.1.2
torchaudio==2.1.2
peft==0.7.1  # For LoRA fine-tuning support

# Google Cloud Text-to-Speech
google-cloud-texttospeech==2.16.4
google-auth==2.25.2
google-oauth2-tool==0.0.3

# Audio processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Data processing and utilities
numpy==1.24.4
pandas==2.1.4
scipy==1.11.4

# Async support
aiofiles==23.2.1

# File handling
python-multipart==0.0.6

# Development and debugging
loguru==0.7.2

# Optional: For better performance
orjson==3.9.10

# Optional: For API documentation
markdown==3.5.1

# File processing for uploads
PyPDF2==3.0.1
pdfplumber==0.9.0
pytesseract==0.3.10
Pillow==10.1.0

# Image generation and RAG dependencies
sentence-transformers==2.2.2
chromadb==0.4.18
svgwrite==1.4.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
